price-compare-mvp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
